requirements.txt
fastapi
uvicorn
python-dotenv
requests
langchain
langchain-community
langchain-openai